+++ src/main.cpp
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   main.cpp                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: emadriga <emadriga@student.42madrid.com>   +#+  +:+       +#+        */
+/*   By: jvacaris <jvacaris@student.42madrid.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/01 17:15:45 by emadriga          #+#    #+#             */
-/*   Updated: 2023/05/23 15:38:23 by emadriga         ###   ########.fr       */
+/*   Updated: 2023/05/24 13:04:48 by jvacaris         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -43,7 +43,7 @@ int main(int argc, char **argv)
 		try
 		{
 			ft::Filetypes aux(NULL);
-			ft::conf newconf("conf/example.conf", aux);
+			ft::conf newconf("conf/jvacaris.conf", aux);
 			int port = 0;
 			if (argc == 3)
 				port = std::atoi(argv[2]);
@@ -82,6 +82,7 @@ int main(int argc, char **argv)
 		ft::Filetypes aux(NULL);
 		ft::conf newconf("conf/example.conf", aux);
 		// std::vector<ft::serverconf>::iterator it = newconf.server;
+		
 
 
 		Request newrequest(file_reader("test_files/get_request", &read_status), newconf.servers[0]);