+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2023-05-24- Jorge
+### Added
+- `Response` is now linked to the server's request reciever.
+
+## 2023-05-23 - Jorge
+### Added
+- We now go through the CGI if appliable.
+
 ## 2023-05-23 - Edu
 ### Added
 - New directory `scripts` to store some bash scripts to automate somw works, including new scripts `fix_header.sh` to set 42 madrid termination mails
@@ -18,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## 2023-05-22 - Edu
 ### Changed
 - Simplifying `server` response mockups
+
 ## 2023-05-22 - Jorge
 ### Added
 - In case a directory is requested and autoindex is disabled, a given index is used.